attempt to restore getAccessModuleConfig() to access-utils
